requirements.txt
asgiref==3.6.0
autopep8==2.0.2
boto3==1.26.114
botocore==1.29.114
Django==4.2
django-cors-headers==3.14.0
django-extensions==3.2.1
django-filter==23.1
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
jmespath==1.0.1
Markdown==3.4.3
mysqlclient==2.1.1
pycodestyle==2.10.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.3
urllib3==1.26.15
